refactor: move similar stuff into their packages
