+++ app/proguard-rules.pro
@@ -13,7 +13,7 @@
                                              # This rule will properly ProGuard all the model classes in
                                              # the package com.yourcompany.models. Modify to fit the structure
                                              # of your app.
-                                             -keepclassmembers class com.amusoft.gdgfirechat.ChatMessage.** {
+                                             -keepclassmembers class com.amusoft.gdgfirechat.model.ChatMessage.** {
                                                *;
                                              }
