--- app/src/main/java/com/amusoft/gdgfirechat/MainActivityKt.kt
@@ -1,12 +1,0 @@
-package com.amusoft.gdgfirechat
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivityKt : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main_kt)
-    }
-}
