+++ app/src/main/res/layout/activity_log_in.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:orientation="vertical"
-    tools:context="com.amusoft.gdgfirechat.LogInActivity">
+    tools:context="com.amusoft.gdgfirechat.ui.LogInActivity">
 
     <TextView
         android:layout_marginTop="60dp"
