+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,6 @@
     android:background="@drawable/background_hd" >
 
     <androidx.appcompat.widget.Toolbar
-        xmlns:android="http://schemas.android.com/apk/res/android"
         android:id="@+id/toolbarAdminDetails"
         android:textColor="@color/white_pure"
         android:layout_width="match_parent"
