+++ app/src/main/res/layout/chat_listitem_left.xml
@@ -14,7 +14,7 @@
         android:orientation="horizontal"
         android:gravity="center_vertical">
 
-        <com.amusoft.gdgfirechat.CircleImageView
+        <com.amusoft.gdgfirechat.view.CircleImageView
             android:id="@+id/image_view"
             android:layout_width="54dp"
             android:layout_height="54dp"
